modified the method and added test conditions

+++ src/main/java/edu/bristol/IMDBRating.java
@@ -5,10 +5,19 @@ public class IMDBRating
     private float currentAverage = 0;
     private int ratingCount = 0;
 
+    // public float addNewRating(int newRating)
+    // {
+    //     int previousTotal = (int) (currentAverage * ratingCount);
+    //     int newTotal = previousTotal + newRating;
+    //     ratingCount++;
+    //     currentAverage = newTotal / ratingCount;
+    //     return currentAverage;
+    // }
+
     public float addNewRating(int newRating)
     {
-        int previousTotal = (int) (currentAverage * ratingCount);
-        int newTotal = previousTotal + newRating;
+        float previousTotal = currentAverage * ratingCount;
+        float newTotal = previousTotal + newRating;
         ratingCount++;
         currentAverage = newTotal / ratingCount;
         return currentAverage;
